refactor json operator modules
